+++ dart/lib/games.dart
@@ -1,1 +1,0 @@
-export 'src/games/tic_tac_toe.dart';